refactor(technologies): restructure pages and add routing
- Split `/technologies` into modular components with a new main, index, and new page structure
- Add `/technologies/new` route to typed-router for future functionality
- Update `/projects/New` route to follow lowercase convention
