+++ frontend/src/pages/technologies/new.vue
@@ -1,0 +1,3 @@
+<template>
+  <div></div>
+</template>
